Restructured the files and added Dockerfile
