+++ domain_app/__init__.py
@@ -1,0 +1,9 @@
+"""
+Author: shaharmelamed
+Date: 3/26/21
+Description: All the routes of the domain_app.
+"""
+# ----- Imports ----- #
+
+from . import api
+from .routes import BLUEPRINT
