+++ domain_app/api/__init__.py
@@ -1,0 +1,8 @@
+"""
+Author: shaharmelamed
+Date: 3/26/21
+Description: The api of the domain_app.
+"""
+# ----- Imports ----- #
+
+from .routes import BLUEPRINT
