+++ static/root_style.css
@@ -9,3 +9,7 @@
 .main {
     text-align: center;
 }
+
+.copyright {
+    font-family: "SF Pro Text", Arial;
+}
